New search controller without api and code cleanup

+++ src/main/java/com/goforcode/grocerygallery/controllers/FridgeController.java
@@ -1,7 +1,6 @@
 package com.goforcode.grocerygallery.controllers;
 
 import java.util.Calendar;
-import java.util.Date;
 import java.util.List;
 
 import org.springframework.security.core.Authentication;
@@ -28,137 +27,53 @@ import com.goforcode.grocerygallery.repositories.ItemRepository;
 public class FridgeController {
 	
 	private ItemRepository itemRepo;
-	private ItemReferenceRepository itemRefRepo; // NEW
+	private ItemReferenceRepository itemRefRepo;
 	
 	public FridgeController(ItemRepository itemRepo, ItemReferenceRepository itemRefRepo) {
 		this.itemRepo = itemRepo;
 		this.itemRefRepo = itemRefRepo;
 	}
 	
-	/* This needs to be fixed before it is uncommented
-	// To allow search to work
-	@GetMapping("")
-	public List<Item> returnItemsInFridge(String partialTitle) {
-		List<Item> returnList;
-		if(partialTitle != null) {
-			returnList = itemRepo.findByTitleContaining(partialTitle);	
-		}
-		else {
-			returnList = itemRepo.findAll();	
-		}
-		return returnList;
-	}*/
-	
 	@GetMapping("")
 	public List<Item> returnItemsInFridge(Authentication auth) {
 		return itemRepo.findByInFridgeTrueAndUserIdEqualsOrderByExpirationDate(getPrincipalUser(auth).getId());
 	}
 	
-	@GetMapping("/search")
-	public List<ItemReference> returnSeachResults(@RequestParam String query) {
-		return itemRefRepo.findByNameLike(query);
-	}
-	
-	
 	@PostMapping("")
 	public Item addItemToFridge(@RequestBody Item fridgeItem, Authentication auth) {
-		// adding completely new food item not in the ref table
-		fridgeItem.setInFridge(true);
-		if (fridgeItem.getId() != null) {
-			Date purchase = fridgeItem.getpurchasedDate();
-			Date expiration = fridgeItem.getExpirationDate();
+		
+		ItemReference itemRef = itemRefRepo.findByNameEquals(fridgeItem.getName());
+				
+		if (itemRef == null || itemRef.getId() == null) {
 			Calendar purchaseDate = Calendar.getInstance();
+			purchaseDate.setTime(fridgeItem.getPurchasedDate());
 			Calendar expirationDate = Calendar.getInstance();
-			purchaseDate.setTime(purchase);
-			expirationDate.setTime(expiration);
-			
-			int daysUntilExpiration = expirationDate.get(Calendar.DAY_OF_YEAR) - purchaseDate.get(Calendar.DAY_OF_YEAR);
-			
-			ItemReference item = new ItemReference();
-			
-			String fridgeItemName = fridgeItem.getName();
-			item.setName(fridgeItemName);
-			item.setShelfLife(daysUntilExpiration);
-			itemRefRepo.save(item);
-			//save item to itemRef table
-			
-			
-			return itemRepo.save(fridgeItem);
-			//save a copy to the item table
-		} else { 
-			//save to the item tables as a new copy
-			ItemReference item = itemRefRepo.findOne(fridgeItem.getId());
-			Item foodItem = new Item();
-			Date purchase = new Date();
-			foodItem.setpurchasedDate(purchase);
-			Calendar purchaseDate = Calendar.getInstance();
-			purchaseDate.setTime(purchase);
-			purchaseDate.add(Calendar.DATE, item.getShelfLife());
-			foodItem.setExpirationDate(purchaseDate.getTime());
-			return itemRepo.save(foodItem);
-			
-			
+			expirationDate.setTime(fridgeItem.getExpirationDate());
+			int shelfLife = expirationDate.get(Calendar.DAY_OF_YEAR) - purchaseDate.get(Calendar.DAY_OF_YEAR);
+			itemRef = new ItemReference(fridgeItem.getName(), shelfLife);
+			itemRefRepo.save(itemRef);
 		}
 		
-//		// create itemrefrepo
-//		
-//		String fridgeItemName = fridgeItem.getName();
-//		List<Item> queryResults = itemRefRepo.findByNameLike(fridgeItemName);
-		
-		
-		
-		// take item in request body, look for it in the reference table by name
-		// if there, calculate default expiration date and set it
-		// save it
-		
-		//if not there, 
-		
-		
-		
-		
-		
-		System.out.println("Item purchase date: " + fridgeItem.getpurchasedDate());
-		System.out.println("Item expiration date: " + fridgeItem.getExpirationDate());
-	
-		
-		
-		//set every new item in fridge not available in other areas
-		fridgeItem.setInGrocery(false);
-		fridgeItem.setWasFinished(false);
-		fridgeItem.setWasWasted(false);
-		
-		//category and date validation if false
-//		fridgeItem.validateCategoryAndDates();
+		fridgeItem.setInFridge(true);
 		fridgeItem.calculateLevel();
-		
 		fridgeItem.setUser(getPrincipalUser(auth));
 		return itemRepo.save(fridgeItem);
 	}
 	
 	@GetMapping("/{id}")
-	public Item getDetailsOfFridgeItem(@PathVariable long id) {
-		Item fridgeItem = itemRepo.findOne(id);
-		return fridgeItem;
+	public Item getDetailsOfFridgeItem(@PathVariable long id, Authentication auth) {
+		return itemRepo.findByIdAndUserId(id, getPrincipalUser(auth).getId());
 	}
 	
 	@PutMapping("/{id}")
 	public Item editFridgeItem(@RequestBody Item fridgeItem, @PathVariable long id, Authentication auth) {
-		
-		//What do we do if this fails (NullPointerException)?
-		User user = (User) auth.getPrincipal();
+		User user = getPrincipalUser(auth);
 		Item searchItem = itemRepo.findByIdAndUserId(id, user.getId());
 		
 		if (searchItem != null) {
 			fridgeItem.setId(id);
 			fridgeItem.setInFridge(true);
 			fridgeItem.setUser(user);
-			
-			//set every updated item in fridge not available in other areas
-			fridgeItem.setInGrocery(false);
-			fridgeItem.setWasFinished(false);
-			fridgeItem.setWasWasted(false);
-			
-//			fridgeItem.validateCategoryAndDates();
 			fridgeItem.calculateLevel();
 			return itemRepo.save(fridgeItem);
 		}
@@ -167,28 +82,25 @@ public class FridgeController {
 	
 	@DeleteMapping("/{id}")
 	public Item deleteFridgeItem(@PathVariable long id, Authentication auth) {
-		User user = (User) auth.getPrincipal();
-		Item fridgeItem = itemRepo.findByIdAndUserId(id, user.getId());
-		if (fridgeItem != null) {
+		Item fridgeItem = itemRepo.findByIdAndUserId(id, getPrincipalUser(auth).getId());
+		
+		if (fridgeItem != null && !fridgeItem.isInGrocery()) {
 			itemRepo.delete(id);
 			return fridgeItem;
+			
+		} else if (fridgeItem != null) {
+			fridgeItem.setInGrocery(true);
+			return itemRepo.save(fridgeItem);
 		}
 		return new Item();
 	}
 	
 	@PostMapping("/{id}/waste")
 	public Item wasteAFridgeItem(@PathVariable long id, Authentication auth) {
-		User user = (User) auth.getPrincipal();
-		Item item = itemRepo.findByIdAndUserId(id, user.getId());
+		Item item = itemRepo.findByIdAndUserId(id, getPrincipalUser(auth).getId());
 		
 		if (item != null) {
 			item.setWasWasted(true);
-			
-			//validation of negative scenarios
-			item.setInFridge(false);
-			item.setInGrocery(false);
-			item.setWasFinished(false);
-			
 			return itemRepo.save(item);
 		}
 		return new Item();
@@ -196,16 +108,9 @@ public class FridgeController {
 	
 	@PostMapping("/{id}/finish")
 	public Item finishAFridgeItem(@PathVariable long id, Authentication auth) {
-		User user = (User) auth.getPrincipal();
-		Item item = itemRepo.findByIdAndUserId(id, user.getId());
+		Item item = itemRepo.findByIdAndUserId(id, getPrincipalUser(auth).getId());
 		if (item != null) {
 			item.setWasFinished(true);
-			
-			//validation of negative scenarios
-			item.setInFridge(false);
-			item.setInGrocery(false);
-			item.setWasWasted(false);
-			
 			return itemRepo.save(item);
 		}
 		return new Item();
@@ -213,23 +118,11 @@ public class FridgeController {
 
 	@PostMapping("/{id}/grocery")
 	public Item moveAFridgeItemToGrocery(@PathVariable long id, @RequestBody Item incomingItem, Authentication auth) {
-		User user = (User) auth.getPrincipal();
-		Item item = itemRepo.findByIdAndUserId(id, user.getId());
+		Item item = itemRepo.findByIdAndUserId(id, getPrincipalUser(auth).getId());
 		
 		if (item != null) {
-			item.setInGrocery(true);
+			item.setInFridgeAndInGrocery();
 			item.setQuantity(incomingItem.getQuantity());
-			
-			//keep it in the fridge
-			item.setInFridge(true);
-			
-			//validation of negative scenarios
-			item.setWasFinished(false);
-			item.setWasWasted(false);
-			
-//			item.validateCategoryAndDates();
-			item.calculateLevel();
-			
 			return itemRepo.save(item);
 		}
 		return new Item();
@@ -237,8 +130,7 @@ public class FridgeController {
 	
 	@GetMapping("count")
 	public int countItemByLevel(@RequestParam int level, Authentication auth) {
-		User user = (User) auth.getPrincipal();
-		return itemRepo.countByInFridgeTrueAndUserIdEqualsAndLevelEquals(user.getId(), level);
+		return itemRepo.countByInFridgeTrueAndUserIdEqualsAndLevelEquals(getPrincipalUser(auth).getId(), level);
 	}
 	
 	public User getPrincipalUser(Authentication auth) {
