+++ src/main/java/com/goforcode/grocerygallery/controllers/GroceryController.java
@@ -41,39 +41,33 @@ public class GroceryController {
 	public Item addItemToGroceryList(@RequestBody Item item, Authentication auth) {
 		User user = (User) auth.getPrincipal();
 		item.setUser(user);
-		item.setInGrocery(true);
-		
-		//validation of negative scenarios
-		item.setInFridge(false);
-		item.setWasFinished(false);
-		item.setWasWasted(false);
-		
+		item.setInGrocery(true);		
 		return itemRepo.save(item);
 	}
 	
 	@GetMapping("/{id}")
-	public Item getDetailsOfGroceryItem(@PathVariable long id) {
-		return itemRepo.findOne(id);
+	public Item getDetailsOfGroceryItem(@PathVariable long id, Authentication auth) {
+		User user = (User) auth.getPrincipal();
+		return itemRepo.findByIdAndUserId(id, user.getId());
 	}
 	
 	@PutMapping("/{id}")
 	public Item editGroceryItem(@PathVariable long id, @RequestBody Item item, Authentication auth) {
-		
-		// What do we do if this errors (NullPointerException)?
 		User user = (User) auth.getPrincipal();
 		Item searchItem = itemRepo.findByIdAndUserId(id, user.getId());
 		
-		if (searchItem != null ) {
+		if (searchItem != null) {
 			
-			item.setInGrocery(true);
+			if (searchItem.isInFridge()) {
+				item.setInFridgeAndInGrocery();
+				item.calculateLevel();
+				
+			} else {
+				item.setInGrocery(true);
+				}
 			item.setUser(user);
 			item.setId(id);
-			
-			//validation of negative scenarios
-			item.setWasFinished(false);
-			item.setWasWasted(false);
-			item.setInFridge(false);
-			return itemRepo.save(item);
+			return itemRepo.save(item);	
 		}
 		return new Item();
 	}
@@ -82,10 +76,15 @@ public class GroceryController {
 	public Item deleteItemFromGroceryList(@PathVariable long id, Authentication auth) {
 		User user = (User) auth.getPrincipal();
 		Item item = itemRepo.findByIdAndUserId(id, user.getId());
-		if (item != null) {
+		if (item != null && !item.isInFridge()) {
 			itemRepo.delete(id);
 			return item;
+			
+		} else if (item !=null) {
+			item.setInFridge(true);
+			return itemRepo.save(item);
 		}
+				
 		return new Item();
 	}
 	
@@ -95,16 +94,8 @@ public class GroceryController {
 		Item item = itemRepo.findByIdAndUserId(id, user.getId());
 		
 		if (item != null) {
-			item.setInFridge(true);
-			
-			//validation of negative scenarios
-			item.setInGrocery(false);
-			item.setWasWasted(false);
-			item.setWasFinished(false);
-			
-//			item.validateCategoryAndDates();
+			item.setInFridgeAndInGrocery();
 			item.calculateLevel();
-			
 			return itemRepo.save(item);
 		}
 		return new Item();
